requirements.txt
streamlit~=1.32.2
pandas~=2.1.0
nltk~=3.8.1
scikit-learn~=1.4.0
numpy~=1.26.0
Pillow~=10.0.0
plotly~=5.18.0